+++ src/Commands/Command.php
@@ -7,6 +7,7 @@ namespace Simtabi\Laranail\Licence\Verifier\Commands;
 use Simtabi\Laranail\Licence\Verifier\LicenceVerifier;
 use Simtabi\Laranail\Licence\Verifier\Contracts\Driver;
 use Simtabi\Laranail\Licence\Verifier\Drivers\DriverManager;
+use Simtabi\Laranail\Package\Tools\Commands\Concerns\ReadsOptions;
 use Simtabi\Laranail\Console\Tools\Commands\Command as ConsoleCommand;
 use Simtabi\Laranail\Console\Tools\Commands\Concerns\SupportsNamespacedNames;
 
@@ -17,6 +18,7 @@ use Simtabi\Laranail\Console\Tools\Commands\Concerns\SupportsNamespacedNames;
  */
 abstract class Command extends ConsoleCommand
 {
+    use ReadsOptions;
     use SupportsNamespacedNames;
 
     protected function manager(): DriverManager